requirements.txt
pycurl
PyYAML
CherryPy
